Replaced linux-dedupe-static.h with a non-copyrightable version
GitHub issue #203 brought up the issue of GPL compliance due to
including a modified Linux kernel header. While this inclusion
should not be an issue under the GPL since APIs are not supposed
to be covered under copyright law, I have further stripped the
static header down to its absolute minimum state and removed the
(not legally applicable) GPL notice from the top.

Anyone who would dispute the claim regarding the GPL not applying
to a bare API header should read this paper before raising the GPL
as an issue for the jdupes project after this change:

https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2893192