Makefile: add USE_JODY_HASH; turn off stack boost with LOW_MEMORY
USE_JODY_HASH has been promoted to a Makefile option. This is a
side effect of making Makefile compile and link ONLY the correct
hash algorithm object files instead of all of them, saving a bit
of wasted binary size in the process.

The problem that originally triggered the stack size increase was
solved when the travdone tree was fixed, but the increase is still
left turned on just in case. Of greater concern is that the stack
size increase is turned on for LOW_MEMORY builds which should not
be asking for 16MiB of stack space, so the Makefile was tweaked to
only add the stack size boost if LOW_MEMORY is not set.